refactor/#42 : usecase 연결 수정, homeview 수정 및 로직 개선

+++ domain/src/main/java/project/graduation/crowd_sourcing/domain/repository/MartSearchRepository.kt
@@ -4,6 +4,6 @@ import project.graduation.crowd_sourcing.domain.model.entity.martsearch.MartEnti
 
 interface MartSearchRepository {
     suspend fun searchMartByZipcode(zipcode: String, radius: Int): List<MartEntity>
-    suspend fun searchMartByLocation(lat: Double, lng: Double, radius: Int): List<MartEntity>
-    suspend fun searchMartByKeyword(keyword: String, radius: Int): List<MartEntity>
+    suspend fun searchMartByLocation(lat: Double, lng: Double, radius: Double): List<MartEntity>
+    suspend fun searchMartByKeyword(keyword: String, radius: Double): List<MartEntity>
 }
